Refactored Event service
